requirements.txt
__future__
abc
collections
dataclasses 
typing
pathlib
gzip
statistics
biopython==1.83
pytest==8.1.1
